+++ app/src/main/java/com/brandonjamesyoung/levelup/fragments/QuestList.kt
@@ -496,6 +496,7 @@ class QuestList: Fragment(R.layout.quest_list) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        InsetHandler.addInsetPadding(requireView())
 
         lifecycleScope.launch(Dispatchers.Main) {
             Log.i(TAG, "On Quest List page")